reorganize files into utils dir
