+++ src/utils/injectHooks.ts
@@ -2,7 +2,7 @@ import type { Config, TypeWithID } from "payload";
 
 import type { PayloadSentinelConfig } from "../config.js";
 
-import { logCollectionAudit, logGlobalAudit } from "../audit/log.js";
+import { logCollectionAudit, logGlobalAudit } from "./log.js";
 
 // Define a type for the options needed by the hook injection
 type HookOptions = Required<
